Crate log4rs_routing_appender

A log4rs appender which routes logging events to dynamically created sub-appenders.

For example, you may want to direct output to different directories based on a "job ID" stored in the MDC:

appenders:
  job:
    kind: routing
    router:
      kind: pattern
      pattern:
        kind: file
        path: "log/jobs/${mdc(job_id)}/output.log"
    cache:
      idle_timeout: 30 seconds
loggers:
  server::job_runner:
    appenders:
    - job

#[macro_use]
extern crate log;
extern crate log_mdc;

let job_id = generate_job_id();
log_mdc::insert("job_id", job_id);

info!("Starting job");Run

Modules

route
Routers.

Structs

RoutingAppender
An appender which routes log events to dynamically constructed sub-appenders.

RoutingAppenderBuilder
A builder for RoutingAppenders.

RoutingAppenderConfig
Configuration for the RoutingAppender.

RoutingAppenderDeserializer
A deserializer for the RoutingAppender.

Functions

register
Registers the following mappings: