Struct log4rs_routing_appender::RoutingAppenderBuilder

pub struct RoutingAppenderBuilder { /* fields omitted */ }

A builder for RoutingAppenders.

Methods

impl RoutingAppenderBuilder

fn idle_timeout(self, idle_timeout: Duration) -> RoutingAppenderBuilder

Sets the duration after which an appender that has not been used will be removed from the cache.

Defaults to 2 minutes.

fn build(self, router: Box<Route>) -> RoutingAppender

Consumes the builder, producing a RoutingAppender.